refactoring searchFragment
